+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2024-06-14
+
+### Changed
+
+- when calibre option is passed now also excludes .calnotes .caltrash
+
+### Added
+
+- hide-dot-files can be passed to ignere .files
+
 ## [1.0.6] - 2024-03-12
 
 ### Security